irrigation/requirements.txt
absl-py
flask
grpcio-tools
kafka-python
mongomock
pymongo
requests
retrying
RPi.GPIO
